+++ ext/ca_obj_unbound_repeat.c
@@ -50,15 +50,9 @@ ca_ubrep_setup (CAUnboundRepeat *ca, CArray *parent,
   ca->rep_ndim  = rep_ndim;
   ca->rep_dim   = ALLOC_N(ca_size_t, rep_ndim);
 
-  memcpy(ca->rep_dim, rep_dim, rep_ndim * sizeof(ca_size_t));
-
-  for (i=0; i<ca->ndim; i++) {
-    if ( ca->rep_dim[i] ) {
-      ca->dim[i] = ca->rep_dim[i];      
-    }
-    else {
-      ca->dim[i] = 1;
-    }
+  for (i=0; i<rep_ndim; i++) {
+    ca->rep_dim[i] = rep_dim[i];
+    ca->dim[i] = ( rep_dim[i] ) ? rep_dim[i] : 1;      
   }
 
   if ( ca_has_mask(parent) ) {
@@ -104,13 +98,6 @@ ca_ubrep_func_ptr_at_addr (void *ap, ca_size_t addr)
   return ca_ptr_at_addr(ca->parent, addr);
 }
 
-static char *
-ca_ubrep_func_ptr_at_index (void *ap, ca_size_t *idx)
-{
-  CAUnboundRepeat *ca = (CAUnboundRepeat *) ap;
-  return ca_ptr_at_index(ca->parent, idx);
-}
-
 static void
 ca_ubrep_func_fetch_addr (void *ap, ca_size_t addr, void *ptr)
 {
@@ -119,24 +106,10 @@ ca_ubrep_func_fetch_addr (void *ap, ca_size_t addr, void *ptr)
 }
 
 static void
-ca_ubrep_func_fetch_index (void *ap, ca_size_t *idx, void *ptr)
-{
-  CAUnboundRepeat *ca = (CAUnboundRepeat *) ap;
-  ca_fetch_index(ca->parent, idx, ptr);
-}
-
-static void
 ca_ubrep_func_store_addr (void *ap, ca_size_t addr, void *ptr)
 {
   CAUnboundRepeat *ca = (CAUnboundRepeat *) ap;
   ca_store_addr(ca->parent, addr, ptr);
-}
-
-static void
-ca_ubrep_func_store_index (void *ap, ca_size_t *idx, void *ptr)
-{
-  CAUnboundRepeat *ca = (CAUnboundRepeat *) ap;
-  ca_store_index(ca->parent, idx, ptr);
 }
 
 static void
@@ -227,11 +200,11 @@ ca_operation_function_t ca_ubrep_func = {
   free_ca_ubrep,
   ca_ubrep_func_clone,
   ca_ubrep_func_ptr_at_addr,
-  ca_ubrep_func_ptr_at_index,
+  NULL, 
   ca_ubrep_func_fetch_addr,
-  ca_ubrep_func_fetch_index,
+  NULL, 
   ca_ubrep_func_store_addr,
-  ca_ubrep_func_store_index,
+  NULL, 
   ca_ubrep_func_allocate,
   ca_ubrep_func_attach,
   ca_ubrep_func_sync,
@@ -304,11 +277,10 @@ rb_ca_unbound_repeat (int argc, VALUE *argv, VALUE self)
   rep_ndim = argc;
 
   count = 0;
-  ndim = 0;
-
+  ndim  = 0;
   elements = 1;
   for (i=0; i<rep_ndim; i++) {
-    if ( rb_obj_is_kind_of(argv[i], rb_cSymbol) ) {
+    if ( TYPE(argv[i]) == T_SYMBOL ) {
       if ( argv[i] == ID2SYM(rb_intern("*")) ) {
         rep_dim[i] = 0;
       }
@@ -321,8 +293,8 @@ rb_ca_unbound_repeat (int argc, VALUE *argv, VALUE self)
         rb_raise(rb_eArgError, "invalid argument");
       }
       rep_dim[i] = ca->dim[count];
-      dim[ndim] = ca->dim[count];
-      elements *= ca->dim[count];
+      dim[ndim]  = ca->dim[count];
+      elements  *= ca->dim[count];
       count++; ndim++;
     }
   }
@@ -334,9 +306,8 @@ rb_ca_unbound_repeat (int argc, VALUE *argv, VALUE self)
   if ( ndim != ca->ndim ) {
     rb_raise(rb_eArgError, "invalid number of nil's (%i for %i)", ndim, ca->ndim);
   }
-  else {
-    return rb_ca_ubrep_new(self, rep_ndim, rep_dim);
-  }
+
+  return rb_ca_ubrep_new(self, rep_ndim, rep_dim);
 }
 
 static VALUE
@@ -388,6 +359,7 @@ ca_ubrep_bind2 (VALUE self, int32_t new_ndim, ca_size_t *new_dim)
   ca_size_t upr_spec[CA_RANK_MAX];
   ca_size_t srp_spec[CA_RANK_MAX];
   int uprep = 0, srp_ndim;
+  int ndim_real;
   int i;
 
   Data_Get_Struct(self, CAUnboundRepeat, ca);
@@ -398,6 +370,7 @@ ca_ubrep_bind2 (VALUE self, int32_t new_ndim, ca_size_t *new_dim)
   }
 
   srp_ndim = 0;
+  ndim_real = 0;
   for (i=0; i<new_ndim; i++) {
     if ( ca->rep_dim[i] == 0 ) {
       if ( new_dim[i] == 0 ) {
@@ -410,14 +383,16 @@ ca_ubrep_bind2 (VALUE self, int32_t new_ndim, ca_size_t *new_dim)
       upr_spec[i] = new_dim[i];
     }
     else {
+      ndim_real++;
       rep_spec[i] = 0;
       srp_spec[srp_ndim++] = 0;
       upr_spec[i] = ca->rep_dim[i];
     }
   }
+  
   if ( uprep ) {
     volatile VALUE rep;
-    if ( srp_ndim >= ca->ndim ) {
+    if ( srp_ndim >= ndim_real ) {
       rep = rb_ca_repeat_new(rb_ca_parent(self), srp_ndim, srp_spec);
     }
     else {
@@ -430,11 +405,8 @@ ca_ubrep_bind2 (VALUE self, int32_t new_ndim, ca_size_t *new_dim)
   }
 }
 
-/* yard:
-  class CAUnboundRepeat
-    def bind_with(other)
-    end
-  end
+/* @overload bind_with (other)
+
 */
 
 VALUE
@@ -460,11 +432,8 @@ ca_ubrep_bind_with (VALUE self, VALUE other)
   }
 }
 
-/* yard:
-  class CAUnboundRepeat
-    def bind(*index)
-    end
-  end
+/* @overload bind (*index)
+
 */
 
 static VALUE