+++ lib/carray/autoload.rb
@@ -3,8 +3,6 @@
 #  carray/autoload.rb
 #
 #  This file is part of Ruby/CArray extension library.
-#  You can redistribute it and/or modify it under the terms of
-#  the Ruby Licence.
 #
 #  Copyright (C) 2005-2010  Hiroki Motoyoshi
 #