+++ lib/carray/inspect.rb
@@ -3,10 +3,8 @@
 #  carray/inspect.rb
 #
 #  This file is part of Ruby/CArray extension library.
-#  You can redistribute it and/or modify it under the terms of
-#  the Ruby Licence.
 #
-#  Copyright (C) 2005 Hiroki Motoyoshi
+#  Copyright (C) 2005-2020 Hiroki Motoyoshi
 #
 # ----------------------------------------------------------------------------
 
@@ -159,7 +157,7 @@ class CArray::Inspector  # :nodoc:
       dim[level].times do |i|
         idx[level] = i
         v = @carray[*idx]
-        if v.undef?
+        if v == UNDEF
           io << '_'
         else
           io << formatter[v]
@@ -178,7 +176,7 @@ class CArray::Inspector  # :nodoc:
       if over
         idx[level] = dim[level] - 1
         v = @carray[*idx]
-        if v.undef?
+        if v == UNDEF
           io << ", _"
         else
           io << ", " + formatter[v]