+++ lib/carray/iterator.rb
@@ -3,10 +3,8 @@
 #  carray/iterator.rb
 #
 #  This file is part of Ruby/CArray extension library.
-#  You can redistribute it and/or modify it under the terms of
-#  the Ruby Licence.
 #
-#  Copyright (C) 2005 Hiroki Motoyoshi
+#  Copyright (C) 2005-2020 Hiroki Motoyoshi
 #
 # ----------------------------------------------------------------------------
 
@@ -318,6 +316,7 @@ class CAClassIterator < CAIterator # :nodoc:
     @reference = reference
     @classifier = classifier || @reference.uniq.sort
     @null = CArray.new(@reference.data_type,[0])
+    @null.data_class = @reference.data_class if @reference.has_data_class?
     @table = {}
     @ndim = 1
     @dim  = [0]